SplitQuery Java client refactor
Several changes:
Update client to match changes on server side
 - splitsPerShard -> splitCount
 - GetMRSplits -> SplitQuery
Move SplitQuery tests from MapReduceIT to VtGateIT
Lower queryserver-config-schema-reload-time to ensure rdonly tablets
update their schema on new DDLs
